Implement Wallet RPC endpoints & add regtest sync
Provide a 'WalletService' trait with implementing struct, and wire it up
to the three preliminary gRPC endpoints: 'WalletBalance', 'NewAddress' &
'ListUnspent'. Also provide a 'WalletService::connect' method to do a
full wallet scan with Bitcoin Core JSON-RPC, against a local bitcoind
instance in regtest mode, using the 'bdk_bitcoind_rpc' crate and adapted
from the example code given on the website:

  https://bitcoindevkit.github.io/book-of-bdk/cookbook/syncing/rpc/

The scan runs just once upon startup of the Rust server, with a hard-
wired data dir: '$PWD/.localnet/bitcoind'. If the sync fails, say
because bitcoind isn't running, the (blocking) thread which performs it
will just panic, for now, but the server will otherwise continue to
function (with an empty wallet, as it is not currently persisted).

(Also fix casing typos & missing utxo 'vout' field in the proto.)

TODO: Add logic to repeatedly sync the wallet.