Add RegisterConfidenceNtfn streaming RPC method
Add a gRPC endpoint, 'Wallet.RegisterConfidenceNtfn', with corresponding
CLI client command, 'notify-confidence', to receive a stream of tx
confirmation events for a given txid. Each event message currently gives
the raw tx bytes, the confirmation status and number (possibly zero) of
confirmations, as well as the hash, height & timestamp of the confirming
block (if any). For now, just a single message is streamed immediately
upon request if the tx is found, with an error produced otherwise.

TODO: Make the stream stay open to send further confidence updates, and
 possibly make it work in the case that the tx hasn't been seen yet, to
 listen for expected publishing events, such as for the deposit tx.