Move confidence streaming to 'wallet.rs' & use missed type alias
Rename 'WalletService::get_tx_confidence' to 'get_tx_confidence_stream',
making it stream a single confidence object, instead of producing the
singleton in 'server.rs'. This is in preparation to adding listener code
to the wallet service so that a continual stream of confidence updates
(tx broadcasts, confirmations, reorgs and conflicts) can be produced.

Also use the predefined 'futures::stream::BoxStream' type alias in place
of the 'Pin<Box<dyn Stream<..> + Send>>' constructs in the server code,
together with the 'StreamExt::boxed' extension method, for simplicity.