Use drop-stream crate to log tx confidence stream closures
Add the small 'drop-stream' dependency to Cargo.toml, which provides a
wrapper struct, 'DropStream', invoking a given callback whenever the
wrapped stream is dropped. This can be used to log both client-side and
server-side closures of a streamed gRPC response.

Use this to print a line to stdout whenever a tx confidence RPC stream
is closed, and with it the attached wallet observer, by wrapping the
pre-boxed return value of 'WalletServiceImpl::get_tx_confidence_stream'.

(TODO: Replace all the println logging with proper logging.)