Undo code workarounds for fixed bugs in Rust IDEA plugin
Remove unnecessary 'std::prelude' imports and uninlined format args
(along with the corresponding Clippy suppression in Cargo.toml). These
were in place to prevent spurious errors from old, buggy versions of the
Intellij Rust plugin, but appear to be fixed at least in the latest
version for IDEA 2024.2. (This is currently maintained in an unofficial
fork:

  https://github.com/l2dy/intellij-rust-unofficial-plugin

since the original open source Rust plugin is deprecated and no longer
available in the marketplace.)

(Also suppress some spell checking warnings for addresses & descriptors,
now that code suppressions for the spell checker work.)