Add Java test client for wallet RPCs & target Java 22
Provide a simple Java test client to query the wallet UTXO set and open
a tx confidence stream for each one concurrently, holding them all open
for 5 seconds before cancelling. This is to better test gRPC server-side
streaming and in particular the behaviour when the stream is closed or
cancelled by either the server or the client.

In order to hopefully simplify the Java code, use blocking gRPC stubs
together with virtual (green) threads, which are available since Java
21, running one-thread-per-task, instead of attempting more complicated
async logic with futures and stream listeners. The former approach ought
to give comparable performance/scalability in production code, so would
likely be worth sticking to for mow.

Also raise the Java source+target version from 11 to 22 in 'pom.xml', to
match that of Bisq2, and add a Maven profile to run the wallet client.