Prevent ConfEvent stream closure with error for unpublished txs
Make the 'walletrpc.ConfEvent.rawTx' proto field optional and add a
'MISSING' confidence type, so that the gRPC client can listen for txs
which have not yet been published when calling 'RegisterConfidenceNtfn'.
In this way, a default "missing tx" confidence event can be streamed if
the tx with the given ID is not yet found in the wallet, and the stream
remains open to await tx broadcast.

Also add a unit test to make sure the correct default is generated by
Prost for 'ConfEvent', in case the proto definitions are altered.