Add integration tests for CLI & make its port configurable
Provide some preliminary Cargo integration tests of the gRPC interface,
which spawn a Wallet RPC service (with an empty wallet) as a Tokio task
and run the 'musig-cli' binary client against it, as a child process.
Currently only the 'wallet-balance' & 'new-address' commands / RPC
endpoints are being tested, with a separate (transient) server created
for each test.

In order to avoid interference between the tests or any real RPC server
running at the time (with subsequent intermittent failures), make each
test server bind to the first free port on localhost starting from 50052
(one higher than the default). This requires the port of the test client
to be configurable, so add a '--port' argument to 'musig-cli' for this
purpose.

(Additionally, fix a Clippy warning in the unit tests that was missed
because they're apparently not linted by default, and require a separate
'--tests' or '--all-targets' argument to "cargo clippy".)